Allow ending copy mode with C-c C-t
When ending julia-vterm-repl-copy-mode by pressing C-c C-t, the Julia
process remains stopped and the buffer is almost unusable, because it
does not have its local keymap set. The only working way to terminate
the copy mode is by pressing RET, but this requires an active region.

This commit fixes the above problem by moving the code for ending the
copy mode to julia-vterm-repl-copy-mode. Now the copy mode can be
terminated by both RET and C-c C-t.